+++ src/Dictionaries/HashedDictionaryCollectionTraits.h
@@ -7,10 +7,7 @@
 #include <Common/HashTable/HashSet.h>
 #include <Common/HashTable/PackedHashMap.h>
 
-namespace DB
-{
-
-namespace HashedDictionaryImpl
+namespace DB::HashedDictionaryImpl
 {
 
 /// sparse_hash_map/sparse_hash_set
@@ -101,7 +98,5 @@ template <typename C> auto resizeContainer(C & c, size_t size) requires (IsBuilt
 template <typename C> void clearContainer(C & c) requires (IsBuiltinHashTable<C>) { return c.clearAndShrink(); }
 
 // NOLINTEND(*)
-
-}
 
 }
