+++ src/Dictionaries/HashedDictionaryCollectionType.h
@@ -9,10 +9,7 @@
 #include <sparsehash/sparse_hash_set>
 #include <type_traits>
 
-namespace DB
-{
-
-namespace HashedDictionaryImpl
+namespace DB::HashedDictionaryImpl
 {
 
 /// Return true if the type is POD [1] for the purpose of layout (this is not
@@ -256,7 +253,5 @@ struct HashedDictionarySetType<dictionary_key_type, /* sparse= */ true, Key>
         HashSet<UInt64, DefaultHash<UInt64>, HashTableGrowerWithPrecalculationAndMaxLoadFactor<>>,
         HashSet<StringRef, DefaultHash<StringRef>, HashTableGrowerWithPrecalculationAndMaxLoadFactor<>>>;
 };
-
-}
 
 }
