+++ utils/check-style/check-large-objects.sh
@@ -10,7 +10,6 @@ TU_EXCLUDES=(
     FunctionsConversion
 
     RangeHashedDictionary
-    HashedDictionary
 
     Aggregator
 )
